fix: explicitly cast int64 to int64 as base `int` type it gets autoassigned could be int32

+++ monero_test.go
@@ -237,7 +237,7 @@ func TestPollMonero(t *testing.T) {
 			"destinations": []map[string]any{
 				{
 					"address": merchantAddr,
-					"amount":  10_000_000_000,
+					"amount":  int64(10_000_000_000),
 				},
 			},
 		},
